requirements.txt
streamlit
python-docx
pandas
openpyxl
google-generativeai